chore: Modify project structure to mirror intended one
