+++ src/app/App.js
@@ -1,0 +1,17 @@
+import React, { Component } from "react";
+import logo from "../assets/logo.svg";
+import "./App.css";
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <header className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+        </header>
+      </div>
+    );
+  }
+}
+
+export default App;
